Flytt synkronisering til eget repo
